+++ libopf_py.pxd
@@ -19,24 +19,24 @@ cdef extern from "subgraph.h":
     void       subgraph_destroy      (subgraph ** sg)
 
     void       subgraph_set_metric   (subgraph *sg,
-                                      float (*arc_weight) (float *f1, float *f2, int n),
+                                      double (*arc_weight) (double *f1, double *f2, int n),
                                       METRIC m)
 
-    bint       subgraph_set_feature  (subgraph *sg, float *feat, int *label, int feat_n)
+    bint       subgraph_set_feature  (subgraph *sg, double *feat, int *label, int feat_n)
 
     bint       subgraph_set_precomputed_distance (subgraph *sg,
-                                                  float *distance,
+                                                  double *distance,
                                                   int *label)
 
     void       subgraph_pdf_evaluate (subgraph * sg)
 
 cdef extern from "supervised.h":
     void opf_supervised_train    (subgraph * sg)
-    void opf_supervised_train_iterative (subgraph *sg, float split)
-    void opf_supervised_train_agglomerative (subgraph *sg, float split)
-    void opf_supervised_classify (subgraph * sgtrain, float *feat, int sample_n, int *label)
+    void opf_supervised_train_iterative (subgraph *sg, double split)
+    void opf_supervised_train_agglomerative (subgraph *sg, double split)
+    void opf_supervised_classify (subgraph * sgtrain, double *feat, int sample_n, int *label)
 
 cdef extern from "unsupervised.h":
     void opf_best_k_min_cut (subgraph * sg, int kmin, int kmax)
     void opf_unsupervised_clustering (subgraph * sg)
-    void opf_unsupervised_knn_classify (subgraph * sgtrain, float *feat, int sample_n, int *label)
+    void opf_unsupervised_knn_classify (subgraph * sgtrain, double *feat, int sample_n, int *label)
